fix: restructure tokens hook
