+++ src/components/SwapWidget/MarketOrder/index.tsx
@@ -9,9 +9,9 @@ import { TRUSTED_TOKEN_ADDRESSES } from 'src/constants/address';
 import { DEFAULT_TOKEN_LISTS_SELECTED } from 'src/constants/lists';
 import { SwapTypes } from 'src/constants/swap';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
-import { useCurrency } from 'src/hooks/Tokens';
 import { MixPanelEvents, useMixpanel } from 'src/hooks/mixpanel';
-import { useTokenHook } from 'src/hooks/multiChainsHooks';
+import { useTokenHook } from 'src/hooks/tokens';
+import { useCurrency } from 'src/hooks/tokens/evm';
 import { useApproveCallbackFromTradeHook } from 'src/hooks/useApproveCallback';
 import { ApprovalState } from 'src/hooks/useApproveCallback/constant';
 import useENS from 'src/hooks/useENS';
