+++ src/components/SwapWidget/SelectTokenDrawer/index.tsx
@@ -6,7 +6,8 @@ import AutoSizer from 'react-virtualized-auto-sizer';
 import { FixedSizeGrid } from 'react-window';
 import Drawer from 'src/components/Drawer';
 import { useChainId } from 'src/hooks';
-import { useAllTokens, useToken } from 'src/hooks/Tokens';
+import { useToken } from 'src/hooks/tokens/evm';
+import { useAllTokens } from 'src/hooks/useAllTokens';
 import usePrevious from 'src/hooks/usePrevious';
 import { useSelectedListInfo } from 'src/state/plists/hooks';
 import { useAddUserToken } from 'src/state/puser/hooks';
