+++ src/hooks/multiChainsHooks.ts
@@ -1,6 +1,5 @@
 /* eslint-disable max-lines */
 import { ChainId } from '@pangolindex/sdk';
-import { useNearToken, useNearTokens, useToken, useTokens, useTokensViaSubGraph } from './Tokens';
 import {
   useDummyTokenCurrencyPrice,
   useDummyTokensCurrencyPrice,
@@ -13,76 +12,6 @@ import {
 export function useDummyHook() {
   return undefined;
 }
-
-export type UseTokenHookType = {
-  [chainId in ChainId]: typeof useToken | typeof useNearToken | typeof useDummyHook;
-};
-
-export const useTokenHook: UseTokenHookType = {
-  [ChainId.FUJI]: useToken,
-  [ChainId.AVALANCHE]: useToken,
-  [ChainId.WAGMI]: useToken,
-  [ChainId.COSTON]: useToken,
-  [ChainId.SONGBIRD]: useToken,
-  [ChainId.FLARE_MAINNET]: useToken,
-  [ChainId.HEDERA_TESTNET]: useToken,
-  [ChainId.HEDERA_MAINNET]: useToken,
-  [ChainId.NEAR_MAINNET]: useNearToken,
-  [ChainId.NEAR_TESTNET]: useNearToken,
-  [ChainId.COSTON2]: useToken,
-  [ChainId.EVMOS_TESTNET]: useToken,
-  [ChainId.EVMOS_MAINNET]: useToken,
-  [ChainId.ETHEREUM]: useDummyHook,
-  [ChainId.POLYGON]: useDummyHook,
-  [ChainId.FANTOM]: useDummyHook,
-  [ChainId.XDAI]: useDummyHook,
-  [ChainId.BSC]: useDummyHook,
-  [ChainId.ARBITRUM]: useDummyHook,
-  [ChainId.CELO]: useDummyHook,
-  [ChainId.OKXCHAIN]: useDummyHook,
-  [ChainId.VELAS]: useDummyHook,
-  [ChainId.AURORA]: useDummyHook,
-  [ChainId.CRONOS]: useDummyHook,
-  [ChainId.FUSE]: useDummyHook,
-  [ChainId.MOONRIVER]: useDummyHook,
-  [ChainId.MOONBEAM]: useDummyHook,
-  [ChainId.OP]: useDummyHook,
-};
-
-export type UseTokensHookType = {
-  [chainId in ChainId]: typeof useTokens | typeof useNearTokens | typeof useTokensViaSubGraph | typeof useDummyHook;
-};
-
-export const useTokensHook: UseTokensHookType = {
-  [ChainId.FUJI]: useTokens,
-  [ChainId.AVALANCHE]: useTokens,
-  [ChainId.WAGMI]: useTokens,
-  [ChainId.COSTON]: useTokens,
-  [ChainId.SONGBIRD]: useTokens,
-  [ChainId.FLARE_MAINNET]: useTokens,
-  [ChainId.HEDERA_TESTNET]: useTokensViaSubGraph,
-  [ChainId.HEDERA_MAINNET]: useTokensViaSubGraph,
-  [ChainId.NEAR_MAINNET]: useNearTokens,
-  [ChainId.NEAR_TESTNET]: useNearTokens,
-  [ChainId.COSTON2]: useTokens,
-  [ChainId.EVMOS_TESTNET]: useTokens,
-  [ChainId.EVMOS_MAINNET]: useTokens,
-  [ChainId.ETHEREUM]: useDummyHook,
-  [ChainId.POLYGON]: useDummyHook,
-  [ChainId.FANTOM]: useDummyHook,
-  [ChainId.XDAI]: useDummyHook,
-  [ChainId.BSC]: useDummyHook,
-  [ChainId.ARBITRUM]: useDummyHook,
-  [ChainId.CELO]: useDummyHook,
-  [ChainId.OKXCHAIN]: useDummyHook,
-  [ChainId.VELAS]: useDummyHook,
-  [ChainId.AURORA]: useDummyHook,
-  [ChainId.CRONOS]: useDummyHook,
-  [ChainId.FUSE]: useDummyHook,
-  [ChainId.MOONRIVER]: useDummyHook,
-  [ChainId.MOONBEAM]: useDummyHook,
-  [ChainId.OP]: useDummyHook,
-};
 
 export type UseTokensCurrencyPriceHookType = {
   [chainId in ChainId]:
