+++ src/hooks/tokens/constant.ts
@@ -1,0 +1,2 @@
+import { Token } from '@pangolindex/sdk';
+export type TokenReturnType = Token | undefined | null;
