+++ src/index.tsx
@@ -10,7 +10,7 @@ import { usePair } from 'src/data/Reserves';
 import { useTotalSupply } from 'src/data/TotalSupply';
 import { useTotalSupplyHook } from 'src/data/multiChainsHooks';
 import { PangolinWeb3Provider, useLibrary } from 'src/hooks';
-import { useAllTokens } from 'src/hooks/Tokens';
+import { useAllTokens } from 'src/hooks/useAllTokens';
 import { useApproveCallbackHook } from 'src/hooks/useApproveCallback';
 import { ApprovalState } from 'src/hooks/useApproveCallback/constant';
 import {
